1219大屏二左侧柱状图四舍五入到百万两位小数

+++ src/main/java/com/spdb/nrpt/entity/report/Retail/RetailReport2PO.java
@@ -14,7 +14,7 @@ import java.util.Date;
 //@JsonSerialize(include=JsonSerialize.Inclusion.NON_NULL)
 public class RetailReport2PO {
 
-    private Long dataValue;
+    private Double dataValue;
 
     private String bankName;
